requirements.txt
python-docx==1.1.2
pydantic==2.11.7
lxml==5.3.0